개발 내용 추가
총알이 발사하는 방향을 바라보도록
적을 랜덤한 위치에서 자동으로 생성되도록
총알과 적이 충돌하면 적의 체력을 소모시키도록(넉백은 추후 구현)

+++ Assets/Scripts/Enemy.cs
@@ -15,7 +15,7 @@ public class Enemy : MonoBehaviour
     public virtual void Move() { }
     public virtual void Attack() { }
 
-    public virtual void Damage(int _damage)
+    public virtual void Damage(float _damage)
     {
         _hp -= _damage;
